pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "verd"
version = "0.2.1"
description = "Multi-LLM debate engine — verdicts everywhere"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Manas Karra"}]
keywords = ["llm", "debate", "verdict", "ai", "multi-model", "code-review"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Quality Assurance",
]
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
slack = [
    "slack-bolt>=1.18.0",
    "httpx>=0.27.0",
]

[project.scripts]
verd      = "verd.__main__:main_default"
verdl     = "verd.__main__:main_light"
verdh     = "verd.__main__:main_heavy"
verd-mcp  = "verd.mcp_server:main"
verd-slack = "verd.slack_bot:main"
